update deno CI
