+++ package.json
@@ -13,7 +13,7 @@
     "count": "tsx script.node/cmd.count.ts",
     "nuke": "tsx script.node/cmd.nuke.ts",
     "mit": "fossa analyze --debug",
-    "d.info": "deno run --allow-read script.deno/info.ts",
+    "d.info": "deno run -R script.deno/info.ts",
     "tmp": "deno run -A script.deno/tmp.ts"
   },
   "workspaces": {
