--- script.deno/build.all.ts
@@ -1,4 +1,0 @@
-/**
- * Bundle modules (ESM)
- */
-console.info('build...🐷 TODO');
