+++ script.deno/info.ts
@@ -1,6 +1,24 @@
+import { Fs, c } from './u.ts';
+
+async function count() {
+  const exclude = [
+    '**/node_modules/**',
+    '**/_archive/**',
+    '**/spikes/**',
+    '**/compiler/**',
+    '**/compiler.samples/**',
+    '**/dist/**',
+  ];
+  const pattern = 'code/**/*.{ts,tsx,mts}';
+  const files = await Fs.glob(import.meta.dirname, '..').find(pattern, { exclude });
+
+  console.info('pattern:  ', c.green(pattern));
+  console.info('files:    ', c.yellow(files.length.toLocaleString()));
+}
+
 /**
  * System/Repo info.
  */
-console.info('↓');
-console.info('hello world 👋');
-import './info.count.ts';
+console.info('↓ 👋');
+await count();
+console.info();
