--- script.deno/yarn.update.ts
@@ -1,29 +1,0 @@
-/**
- * TODO 🐷 WIP
- */
-console.log('🐷 WIP', 'update yarn references');
-
-//
-// const process = Deno.run({
-//   cmd: ['yarn', 'outdated'],
-//   stdout: 'piped', // Capture standard output
-//   stderr: 'piped', // Capture standard error
-// });
-//
-// // console.log('res', res);
-// // Await the process to finish
-// const { code } = await process.status();
-//
-// // Checking if the process exited successfully
-// if (code === 0) {
-//   const rawOutput = await process.output();
-//   const outputString = new TextDecoder().decode(rawOutput);
-//   console.log('Success:', outputString);
-// } else {
-//   const rawError = await process.stderrOutput();
-//   const errorString = new TextDecoder().decode(rawError);
-//   console.error('Failed with error:', errorString);
-// }
-//
-// // Always close the process
-// process.close();
